backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
requests
mangum
strands-agents
python-jose[cryptography]
pyjwt[crypto]
anthropic
openai>=1.0.0
boto3
botocore
